+++ services/explorer-api/src/event-handler/on-pending-txs.ts
@@ -1,12 +1,12 @@
 import { PendingTxsEvent } from "@chicmoz-pkg/message-registry";
 import { chicmozL2PendingTxSchema } from "@chicmoz-pkg/types";
 import { logger } from "../logger.js";
+import { storeL2Tx } from "../database/controllers/l2contract/store.js";
 
-// eslint-disable-next-line @typescript-eslint/require-await
 export const onPendingTxs = async ({ txs }: PendingTxsEvent) => {
-  logger.info(`🔗 Pending txs: ${txs.length}`);
   for (const tx of txs) {
+    logger.info(`🔗 Pending tx: ${tx.hash}`);
     const res = chicmozL2PendingTxSchema.parse(tx);
-    logger.info(res.hash);
+    await storeL2Tx(res);
   }
 };
